% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{AA2DACOR}
\alias{AA2DACOR}
\title{2D Autocorrelations Descriptors for 20 Amino Acids calculated by Dragon}
\usage{
data(AA2DACOR)
}
\description{
This dataset includes the 2D autocorrelations descriptors of
the 20 amino acids calculated by Dragon (version 5.4)
used for scales extraction in this package.
}
\examples{
data(AA2DACOR)

}
\keyword{datasets}

