\docType{data}
\name{AA3DMoRSE}
\alias{AA3DMoRSE}
\title{3D-MoRSE Descriptors for 20 Amino Acids calculated by Dragon}
\usage{
data(AA3DMoRSE)
}
\description{
3D-MoRSE Descriptors for 20 Amino Acids calculated by
Dragon
}
\details{
This dataset includes the 3D-MoRSE descriptors of the 20
amino acids calculated by Dragon (version 5.4) used for
scales extraction in this package.
}
\examples{
data(AA3DMoRSE)
}
\keyword{AA3DMoRSE}

