\docType{data}
\name{AAACF}
\alias{AAACF}
\title{Atom-Centred Fragments Descriptors for 20 Amino Acids calculated by Dragon}
\usage{
data(AAACF)
}
\description{
Atom-Centred Fragments Descriptors for 20 Amino Acids
calculated by Dragon
}
\details{
This dataset includes the atom-centred fragments
descriptors of the 20 amino acids calculated by Dragon
(version 5.4) used for scales extraction in this package.
}
\examples{
data(AAACF)
}
\keyword{AAACF}

