\docType{data}
\name{AABLOSUM45}
\alias{AABLOSUM45}
\title{BLOSUM45 Matrix for 20 Amino Acids}
\usage{
data(AABLOSUM45)
}
\description{
BLOSUM45 Matrix for 20 Amino Acids
}
\details{
BLOSUM45 Matrix for the 20 amino acids. The matrix was
extracted from the \code{Biostrings} package of
Bioconductor.
}
\examples{
data(AABLOSUM45)
}
\keyword{AABLOSUM45}
\keyword{BLOSUM}
\keyword{datasets}

