% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{AABLOSUM62}
\alias{AABLOSUM62}
\title{BLOSUM62 Matrix for 20 Amino Acids}
\usage{
data(AABLOSUM62)
}
\description{
BLOSUM62 Matrix for 20 Amino Acids
}
\details{
BLOSUM62 Matrix for the 20 amino acids. The matrix was extracted from the
\code{Biostrings} package of Bioconductor.
}
\examples{
data(AABLOSUM62)
}
\keyword{AABLOSUM62}
\keyword{BLOSUM}
\keyword{datasets}

