\docType{data}
\name{AAConn}
\alias{AAConn}
\title{Connectivity Indices Descriptors for 20 Amino Acids calculated by Dragon}
\usage{
data(AAConn)
}
\description{
Connectivity Indices Descriptors for 20 Amino Acids
calculated by Dragon
}
\details{
This dataset includes the connectivity indices descriptors
of the 20 amino acids calculated by Dragon (version 5.4)
used for scales extraction in this package.
}
\examples{
data(AAConn)
}
\keyword{AAConn}

