\docType{data}
\name{AADescAll}
\alias{AADescAll}
\title{All 2D Descriptors for 20 Amino Acids calculated by Dragon}
\description{
  All 2D Descriptors for 20 Amino Acids calculated by
  Dragon
}
\details{
  This dataset includes all the 2D descriptors of the 20
  amino acids calculated by Dragon (version 5.4) used for
  scales extraction in this package.
}
\examples{
data(AADescAll)
}
\keyword{AADescAll}
\keyword{datasets}

