\docType{data}
\name{AAEdgeAdj}
\alias{AAEdgeAdj}
\title{Edge Adjacency Indices Descriptors for 20 Amino Acids calculated by Dragon}
\description{
  Edge Adjacency Indices Descriptors for 20 Amino Acids
  calculated by Dragon
}
\details{
  This dataset includes the edge adjacency indices
  descriptors of the 20 amino acids calculated by Dragon
  (version 5.4) used for scales extraction in this package.
}
\examples{
data(AAEdgeAdj)
}
\keyword{AAEdgeAdj}
\keyword{datasets}

