\docType{data}
\name{AAGeom}
\alias{AAGeom}
\title{Geometrical Descriptors for 20 Amino Acids calculated by Dragon}
\description{
  Geometrical Descriptors for 20 Amino Acids calculated by
  Dragon
}
\details{
  This dataset includes the geometrical descriptors of the
  20 amino acids calculated by Dragon (version 5.4) used
  for scales extraction in this package.
}
\examples{
data(AAGeom)
}
\keyword{AAGeom}
\keyword{datasets}

