% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{AAindex}
\alias{AAindex}
\title{AAindex Data of 544 Physicochemical and Biological Properties
for 20 Amino Acids}
\description{
The data was extracted from the AAindex1 database ver 9.1
(\url{https://www.genome.jp/ftp/db/community/aaindex/old/ver9.1/aaindex1})
as of November, 2012 (data last modified on 2006-08-14).
}
\details{
With this dataset, users can investigate each property's accession number
and other details. See \url{https://www.genome.jp/aaindex/}
for more information.
}
\examples{
data(AAindex)
}
