% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{AAInfo}
\alias{AAInfo}
\title{Information Indices Descriptors for 20 Amino Acids calculated by Dragon}
\description{
This dataset includes the information indices descriptors of
the 20 amino acids calculated by Dragon (version 5.4)
used for scales extraction in this package.
}
\examples{
data(AAInfo)

}
\keyword{datasets}
