% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{AAMetaInfo}
\alias{AAMetaInfo}
\title{Meta Information for the 20 Amino Acids}
\description{
This dataset includes the meta information of
the 20 amino acids used for the 2D and 3D descriptor
calculation in this package. Each column represents:
\describe{
  \item{\code{AAName}}{Amino acid name}
  \item{\code{Short}}{One-letter representation}
  \item{\code{Abbreviation}}{Three-letter representation}
  \item{\code{mol}}{SMILES representation}
  \item{\code{PUBCHEM_COMPOUND_CID}}{PubChem CID for the amino acid}
  \item{\code{PUBCHEM_LINK}}{PubChem link for the amino acid}
}
}
\examples{
data(AAMetaInfo)
}
