% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{AAMOE2D}
\alias{AAMOE2D}
\title{2D Descriptors for 20 Amino Acids calculated by MOE 2011.10}
\usage{
data(AAMOE2D)
}
\description{
2D Descriptors for 20 Amino Acids calculated by MOE 2011.10
}
\details{
This dataset includes the 2D descriptors of
the 20 amino acids calculated by MOE 2011.10
used for scales extraction in this package.
}
\examples{
data(AAMOE2D)
}
\keyword{datasets}

