\docType{data}
\name{AAMolProp}
\alias{AAMolProp}
\title{Molecular Properties Descriptors for 20 Amino Acids calculated by Dragon}
\usage{
data(AAMolProp)
}
\description{
Molecular Properties Descriptors for 20 Amino Acids
calculated by Dragon
}
\details{
This dataset includes the molecular properties descriptors
of the 20 amino acids calculated by Dragon (version 5.4)
used for scales extraction in this package.
}
\examples{
data(AAMolProp)
}
\keyword{AAMolProp}

