% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{AAPAM120}
\alias{AAPAM120}
\title{PAM120 Matrix for 20 Amino Acids}
\description{
PAM120 Matrix for the 20 amino acids. The matrix was extracted from the
\code{Biostrings} package of Bioconductor.
}
\examples{
data(AAPAM120)
}
