\docType{data}
\name{AAPAM70}
\alias{AAPAM70}
\title{PAM70 Matrix for 20 Amino Acids}
\description{
  PAM70 Matrix for 20 Amino Acids
}
\details{
  PAM70 Matrix for the 20 amino acids. The matrix was
  extracted from the \code{Biostrings} package of
  Bioconductor.
}
\examples{
data(AAPAM70)
}
\keyword{AAPAM70}
\keyword{PAM}
\keyword{datasets}

