\docType{data}
\name{AARDF}
\alias{AARDF}
\title{RDF Descriptors for 20 Amino Acids calculated by Dragon}
\usage{
data(AARDF)
}
\description{
RDF Descriptors for 20 Amino Acids calculated by Dragon
}
\details{
This dataset includes the RDF descriptors of the 20 amino
acids calculated by Dragon (version 5.4) used for scales
extraction in this package.
}
\examples{
data(AARDF)
}
\keyword{AARDF}

