\docType{data}
\name{AATopoChg}
\alias{AATopoChg}
\title{Topological Charge Indices Descriptors for 20 Amino Acids calculated by Dragon}
\description{
  Topological Charge Indices Descriptors for 20 Amino Acids
  calculated by Dragon
}
\details{
  This dataset includes the topological charge indices
  descriptors of the 20 amino acids calculated by Dragon
  (version 5.4) used for scales extraction in this package.
}
\examples{
data(AATopoChg)
}
\keyword{AATopoChg}
\keyword{datasets}

