% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extractFAScales}
\alias{extractFAScales}
\title{Generalized Scales-Based Descriptors derived by Factor Analysis}
\usage{
extractFAScales(x, propmat, factors, scores = "regression", lag,
  scale = TRUE, silent = TRUE)
}
\arguments{
\item{x}{A character vector, as the input protein sequence.}

\item{propmat}{A matrix containing the properties for the amino acids.
Each row represent one amino acid type, each column represents one property.
Note that the one-letter row names must be provided for we need them to seek
the properties for each AA type.}

\item{factors}{Integer. The number of factors to be fitted.
Must be no greater than the number of AA properties provided.}

\item{scores}{Type of scores to produce. The default is \code{"regression"},
which gives Thompson's scores, \code{"Bartlett"} given Bartlett's weighted
least-squares scores.}

\item{lag}{The lag parameter. Must be less than the amino acids number
in the protein sequence.}

\item{scale}{Logical. Should we auto-scale the property matrix
(\code{propmat}) before doing Factor Analysis? Default is \code{TRUE}.}

\item{silent}{Logical. Whether we print the SS loadings,
       proportion of variance and the cumulative proportion of
       the selected factors or not.
       Default is \code{TRUE}.}
}
\value{
A length \code{lag * p^2} named vector,
        \code{p} is the number of scales (factors) selected.
}
\description{
Generalized Scales-Based Descriptors derived by Factor Analysis
}
\details{
This function calculates the generalized scales-based descriptors
derived by Factor Analysis (FA).
Users could provide customized amino acid property matrices.
}
\examples{
x = readFASTA(system.file('protseq/P00750.fasta', package = 'protr'))[[1]]
data(AATopo)
tprops = AATopo[, c(37:41, 43:47)]  # select a set of topological descriptors
fa = extractFAScales(x, propmat = tprops, factors = 5, lag = 7, silent = FALSE)
}
\author{
Nan Xiao <\url{http://r2s.name}>
}
\references{
Atchley, W. R., Zhao, J., Fernandes, A. D., & Druke, T. (2005).
Solving the protein sequence metric problem.
Proceedings of the National Academy of Sciences of the United States of America,
102(18), 6395-6400.
}
\keyword{Analysis}
\keyword{Factor}
\keyword{PCM}
\keyword{extract}
\keyword{extractFAScales}

