% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pcm01extractScalesGap.R
\name{extractScalesGap}
\alias{extractScalesGap}
\title{Scales-Based Descriptors derived by Principal Components Analysis
(with Gap Support)}
\usage{
extractScalesGap(x, propmat, pc, lag, scale = TRUE, silent = TRUE)
}
\arguments{
\item{x}{A character vector, as the input protein sequence.
Use '\code{-}' to represent gaps in the sequence.}

\item{propmat}{A matrix containing the properties for the amino acids.
Each row represent one amino acid type, each column represents
one property. Note that the one-letter row names must be provided
for we need them to seek the properties for each AA type.}

\item{pc}{Integer. Use the first pc principal components as the scales.
Must be no greater than the number of AA properties provided.}

\item{lag}{The lag parameter. Must be less than the amino acids.}

\item{scale}{Logical. Should we auto-scale the property matrix
(\code{propmat}) before PCA? Default is \code{TRUE}.}

\item{silent}{Logical. Whether we print the standard deviation,
       proportion of variance and the cumulative proportion of
       the selected principal components or not.
       Default is \code{TRUE}.}
}
\value{
A length \code{lag * p^2} named vector,
        \code{p} is the number of scales (principal components) selected.
}
\description{
Scales-Based Descriptors derived by Principal Components Analysis
(with Gap Support)
}
\details{
This function calculates scales-based descriptors
derived by Principal Components Analysis (PCA), with gap support.
Users could provide customized amino acid property matrices.
This function implements the core computation procedure needed for
the scales-based descriptors derived by AA-Properties (AAindex)
and scales-based descriptors derived by 20+ classes of 2D and 3D
molecular descriptors (Topological, WHIM, VHSE, etc.) in the protr package.
}
\examples{
# amino acid sequence with gaps
x = readFASTA(system.file('protseq/align.fasta', package = 'protr'))$`IXI_235`
data(AAindex)
AAidxmat = t(na.omit(as.matrix(AAindex[, 7:26])))
scales = extractScalesGap(x, propmat = AAidxmat, pc = 5, lag = 7, silent = FALSE)
}
\author{
Nan Xiao <\url{http://r2s.name}>
}
\seealso{
See \code{\link{extractProtFPGap}} for amino acid property based
scales descriptors (protein fingerprint) with gap support.
}
\keyword{PCA}
\keyword{PCM}
\keyword{extract}
\keyword{gap}
\keyword{scales}

