% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-05-getUniProt.R
\name{getUniProt}
\alias{getUniProt}
\title{Retrieve Protein Sequences from UniProt by Protein ID}
\usage{
getUniProt(id)
}
\arguments{
\item{id}{A character vector, as the protein ID(s).}
}
\value{
A list, each component contains one protein sequence.
}
\description{
This function retrieves protein sequences from uniprot.org by protein ID(s).
}
\examples{
\dontrun{
# Network latency may slow down this example
# Only test this when your connection is fast enough
ids <- c("P00750", "P00751", "P00752")
getUniProt(ids)
}
}
\seealso{
See \code{\link{readFASTA}} for reading FASTA format files.
}
\author{
Nan Xiao <\url{https://nanx.me}>
}
