\docType{data}
\name{OptAA3d}
\alias{OptAA3d}
\title{OptAA3d.sdf semiempirical AM1}
\description{
  OptAA3d.sdf semiempirical AM1
}
\details{
  OptAA3d.sdf semiempirical AM1
}
\examples{
# This operation requires the rcdk package
# require(rcdk)
# read.csv(OptAA3d.sdf, package = 'protr')
}
\keyword{Acid}
\keyword{Amino}
\keyword{OptAA3d}
\keyword{datasets}

