\docType{package}
\name{rdpi-package}
\alias{rdpi}
\alias{rdpi-package}
\title{Protein Sequence Feature Extraction with R}
\description{
  The rdpi package aims for protein sequence feature
  extraction, which could be easily applied in
  Chemoinformatics, Bioinforamtics and Chemogenomics
  research. The features listed in this package include
  Amino Acid Composition (Amino Acid Composition/Dipeptide
  Composition/Tripeptide Composition), Autocorrelation
  (Normalized Moreau-Broto Autocorrelation/Moran
  Autocorrelation/Geary Autocorrelation), CTD
  (Composition/Transition/Distribution), Conjoint Traid,
  Quasi-sequence Order (Sequence Order Coupling
  Number/Quasi-sequence Order Descriptors), and Pseudo
  Amino Acid Composition (Pseudo Amino Acid
  Composition/Amphiphilic Pseudo Amino Acid Composition).
  The package is developed by Computational Biology and
  Drug Design Group, Central South University.
}
\details{
  \tabular{ll}{ Package: \tab rdpi\cr Type: \tab Package\cr
  Version: \tab 0.1-0\cr License: \tab BSD\cr }
}
\note{
  Bug reports and feature requests should be sent to
  \url{https://github.com/road2stat/rdpi/issues}.
}
\section{Amino Acid Composition}{
  put AAC description here.
}

\section{Dipeptide Composition}{
  put DC description here.
}
\examples{

}
\author{
  Xiao Nan <\url{http://www.road2stat.com}> Dongsheng Cao
  \email{oriental-cds@163.com} Qingsong Xu
  \email{dasongxu@gmail.com} Yizeng Liang
  \email{yizeng_liang@263.net}
}
\references{
  The associated website for this package:
  \url{http://cbdd.csu.edu.cn/rdpi/}
}
\seealso{
  Add \code{\link{rca}}, \code{\link{lfda}},
  \code{\link{klfda}}, \code{\link{dca}},
  \code{\link{kdca}}, etc.
}
\keyword{acid}
\keyword{amino}
\keyword{bioinforamtics}
\keyword{chemogenomics}
\keyword{chemoinformatics}
\keyword{descriptors}
\keyword{extraction}
\keyword{feature}
\keyword{package}
\keyword{protein}
\keyword{sequence}

