% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{AABLOSUM100}
\alias{AABLOSUM100}
\title{BLOSUM100 Matrix for 20 Amino Acids}
\usage{
data(AABLOSUM100)
}
\description{
BLOSUM100 Matrix for 20 Amino Acids
}
\details{
BLOSUM100 Matrix for the 20 amino acids. The matrix was extracted from the
\code{Biostrings} package of Bioconductor.
}
\examples{
data(AABLOSUM100)
}
\keyword{AABLOSUM100}
\keyword{BLOSUM}
\keyword{datasets}

