% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{AABLOSUM100}
\alias{AABLOSUM100}
\title{BLOSUM100 Matrix for 20 Amino Acids}
\usage{
data(AABLOSUM100)
}
\description{
BLOSUM100 Matrix for 20 Amino Acids
}
\details{
BLOSUM100 Matrix for the 20 amino acids. The matrix was extracted from the
\code{Biostrings} package of Bioconductor.
}
\examples{
data(AABLOSUM100)
}
\keyword{datasets}

