% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{AABurden}
\alias{AABurden}
\title{Burden Eigenvalues Descriptors for 20 Amino Acids calculated by Dragon}
\usage{
data(AABurden)
}
\description{
Burden Eigenvalues Descriptors for 20 Amino Acids calculated by Dragon
}
\details{
This dataset includes the Burden eigenvalues descriptors of
the 20 amino acids calculated by Dragon (version 5.4)
used for scales extraction in this package.
}
\examples{
data(AABurden)
}
\keyword{datasets}

