% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{AAConn}
\alias{AAConn}
\title{Connectivity Indices Descriptors for 20 Amino Acids calculated by Dragon}
\usage{
data(AAConn)
}
\description{
This dataset includes the connectivity indices descriptors of
the 20 amino acids calculated by Dragon (version 5.4)
used for scales extraction in this package.
}
\examples{
data(AAConn)

}
\keyword{datasets}

