% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{AACPSA}
\alias{AACPSA}
\title{CPSA Descriptors for 20 Amino Acids calculated by Discovery Studio}
\usage{
data(AACPSA)
}
\description{
This dataset includes the CPSA descriptors of the 20 amino acids
calculated by Discovery Studio (version 2.5) used for scales extraction
in this package.
}
\details{
All amino acid molecules had also been optimized with MOE 2011.10
(semiempirical AM1) before calculating these CPSA descriptors.
The SDF file containing the information of the optimized amino acid
molecules is included in this package. See \code{\link{OptAA3d}}
for more information.
}
\examples{
data(AACPSA)

}
\keyword{datasets}

