% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{AADescAll}
\alias{AADescAll}
\title{All 2D Descriptors for 20 Amino Acids calculated by Dragon}
\description{
This dataset includes all the 2D descriptors of
the 20 amino acids calculated by Dragon (version 5.4)
used for scales extraction in this package.
}
\examples{
data(AADescAll)
}
