% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{AAEdgeAdj}
\alias{AAEdgeAdj}
\title{Edge Adjacency Indices Descriptors for 20 Amino Acids calculated by Dragon}
\usage{
data(AAEdgeAdj)
}
\description{
This dataset includes the edge adjacency indices descriptors of
the 20 amino acids calculated by Dragon (version 5.4)
used for scales extraction in this package.
}
\examples{
data(AAEdgeAdj)

}
\keyword{datasets}
