% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{AAFGC}
\alias{AAFGC}
\title{Functional Group Counts Descriptors for 20 Amino Acids calculated by Dragon}
\usage{
data(AAFGC)
}
\description{
Functional Group Counts Descriptors for 20 Amino Acids calculated by Dragon
}
\details{
This dataset includes the functional group counts descriptors of
the 20 amino acids calculated by Dragon (version 5.4)
used for scales extraction in this package.
}
\examples{
data(AAFGC)
}
\keyword{AAFGC}

