% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{AAGeom}
\alias{AAGeom}
\title{Geometrical Descriptors for 20 Amino Acids calculated by Dragon}
\usage{
data(AAGeom)
}
\description{
Geometrical Descriptors for 20 Amino Acids calculated by Dragon
}
\details{
This dataset includes the geometrical descriptors of
the 20 amino acids calculated by Dragon (version 5.4)
used for scales extraction in this package.
}
\examples{
data(AAGeom)

}
\keyword{datasets}

