\docType{data}
\name{AAGETAWAY}
\alias{AAGETAWAY}
\title{GETAWAY Descriptors for 20 Amino Acids calculated by Dragon}
\description{
  GETAWAY Descriptors for 20 Amino Acids calculated by
  Dragon
}
\details{
  This dataset includes the GETAWAY descriptors of the 20
  amino acids calculated by Dragon (version 5.4) used for
  scales extraction in this package.
}
\examples{
data(AAGETAWAY)
}
\keyword{AAGETAWAY}
\keyword{datasets}

