% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{AAindex}
\alias{AAindex}
\title{AAindex Data of 544 Physicochemical and Biological Properties
for 20 Amino Acids}
\usage{
data(AAindex)
}
\description{
The data was extracted from the AAindex1 database ver 9.1
(\url{ftp://ftp.genome.jp/pub/db/community/aaindex/aaindex1}) as of
November, 2012 (Data Last Modified 2006-08-14).
}
\details{
With this dataset, users can investigate each property's accession number
and other details. Visit \url{http://www.genome.jp/dbget/aaindex.html}
for more information.
}
\examples{
data(AAindex)

}
\keyword{AAindex}
\keyword{datasets}
