\docType{data}
\name{AAindex}
\alias{AAindex}
\title{AAindex Data of 544 Physicochemical and Biological Properties
for 20 Amino Acids}
\description{
  AAindex Data of 544 Physicochemical and Biological
  Properties for 20 Amino Acids
}
\details{
  The data was extracted from the AAindex1 database ver 9.1
  (\url{ftp://ftp.genome.jp/pub/db/community/aaindex/aaindex1})
  as of Nov. 2012 (Data Last Modified 2006-08-14). Visit
  \url{http://www.genome.jp/dbget/aaindex.html} for more
  details.
}
\examples{
data(AAindex)
}
\keyword{aaindex}
\keyword{AAindex}
\keyword{datasets}

