\docType{data}
\name{AAPAM120}
\alias{AAPAM120}
\title{PAM120 Matrix for 20 Amino Acids}
\usage{
data(AAPAM120)
}
\description{
PAM120 Matrix for 20 Amino Acids
}
\details{
PAM120 Matrix for the 20 amino acids. The matrix was
extracted from the \code{Biostrings} package of
Bioconductor.
}
\examples{
data(AAPAM120)
}
\keyword{AAPAM120}
\keyword{PAM}
\keyword{datasets}

