% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{AAPAM250}
\alias{AAPAM250}
\title{PAM250 Matrix for 20 Amino Acids}
\usage{
data(AAPAM250)
}
\description{
PAM250 Matrix for 20 Amino Acids
}
\details{
PAM250 Matrix for the 20 amino acids. The matrix was extracted from the
\code{Biostrings} package of Bioconductor.
}
\examples{
data(AAPAM250)
}
\keyword{AAPAM250}
\keyword{PAM}
\keyword{datasets}

