% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{AAPAM250}
\alias{AAPAM250}
\title{PAM250 Matrix for 20 Amino Acids}
\usage{
data(AAPAM250)
}
\description{
PAM250 Matrix for 20 Amino Acids
}
\details{
PAM250 Matrix for the 20 amino acids. The matrix was extracted from the
\code{Biostrings} package of Bioconductor.
}
\examples{
data(AAPAM250)
}
\keyword{datasets}

