% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{AAPAM30}
\alias{AAPAM30}
\title{PAM30 Matrix for 20 Amino Acids}
\usage{
data(AAPAM30)
}
\description{
PAM30 Matrix for the 20 amino acids. The matrix was extracted from the
\code{Biostrings} package of Bioconductor.
}
\examples{
data(AAPAM30)

}
\keyword{datasets}
