% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{AAPAM70}
\alias{AAPAM70}
\title{PAM70 Matrix for 20 Amino Acids}
\usage{
data(AAPAM70)
}
\description{
PAM70 Matrix for the 20 amino acids. The matrix was extracted from the
\code{Biostrings} package of Bioconductor.
}
\examples{
data(AAPAM70)

}
\keyword{datasets}

