% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{AARandic}
\alias{AARandic}
\title{Randic Molecular Profiles Descriptors for 20 Amino Acids calculated by Dragon}
\usage{
data(AARandic)
}
\description{
This dataset includes the Randic molecular profiles descriptors of
the 20 amino acids calculated by Dragon (version 5.4)
used for scales extraction in this package.
}
\examples{
data(AARandic)

}
\keyword{datasets}

