% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{AAWHIM}
\alias{AAWHIM}
\title{WHIM Descriptors for 20 Amino Acids calculated by Dragon}
\usage{
data(AAWHIM)
}
\description{
This dataset includes the WHIM descriptors of
the 20 amino acids calculated by Dragon (version 5.4)
used for scales extraction in this package.
}
\examples{
data(AAWHIM)

}
\keyword{datasets}
