% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protr-datalist.R
\docType{data}
\name{OptAA3d}
\alias{OptAA3d}
\title{OptAA3d.sdf - 20 Amino Acids Optimized with MOE 2011.10 (Semiempirical AM1)}
\description{
OptAA3d.sdf - 20 Amino Acids Optimized with MOE 2011.10 (Semiempirical AM1)
}
\examples{
# This operation requires the rcdk package
# require(rcdk)
# optaa3d = load.molecules(system.file("sysdata/OptAA3d.sdf", package = "protr"))
# view.molecule.2d(optaa3d[[1]])  # view the first AA
}
\keyword{datasets}
