% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/protr-package.R
\docType{package}
\name{protr-package}
\alias{protr}
\alias{protr-package}
\title{Generating Various Numerical Representation Schemes of Protein Sequence}
\description{
The protr package is a comprehensive toolkit for
generating various numerical representation schemes of protein sequence.
The descriptors are extensively utilized in bioinformatics and
chemogenomics research. The commonly used descriptors include amino acid
composition, autocorrelation, CTD, conjoint traid, quasi-sequence order,
pseudo amino acid composition, and profile-based descriptors derived by
Position-Specific Scoring Matrix (PSSM). The descriptors for
proteochemometric (PCM) modeling include the scales-based
descriptors derived by principal components analysis, factor analysis,
multidimensional scaling, amino acid properties (AAindex), 20+ classes of
2D and 3D molecular descriptors (Topological, WHIM, VHSE, etc.), and
BLOSUM/PAM matrix-derived descriptors. The protr package also
integrates the function of parallelized similarity computation derived by
pairwise protein sequence alignment and Gene Ontology (GO) semantic
similarity measures.
}
\details{
\tabular{ll}{ Package: \tab protr\cr Type: \tab Package\cr
Version: \tab 0.5-1\cr License: \tab BSD 3-clause License\cr }
}
\note{
The comprehensive user guide could be opened with \code{vignette('protr')},
which explains each descriptor included in this package and corresponding
usage.

The web server for this package, \code{ProtrWeb} is located at:
\url{http://cbdd.csu.edu.cn:8080/protrweb/}.

Bug reports and feature requests should be sent to
\url{https://github.com/road2stat/protr/issues}.
}
\examples{
NULL
}
\author{
Nan Xiao <\email{road2stat@gmail.com}>
        Qing-Song Xu <\email{dasongxu@gmail.com}>
        Dong-Sheng Cao <\email{oriental-cds@163.com}>
}
\references{
(to appear)
}

