% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protr-package.R
\docType{package}
\name{protr-package}
\alias{protr}
\alias{protr-package}
\title{protr: Generating Various Numerical Representation Schemes for Protein Sequences}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Comprehensive toolkit for generating various numerical features of protein sequences described in Xiao et al. (2015) \doi{10.1093/bioinformatics/btv042}. For full functionality, the software 'ncbi-blast+' is needed, see \url{https://blast.ncbi.nlm.nih.gov/Blast.cgi?PAGE_TYPE=BlastDocs&DOC_TYPE=Download} for more information.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nanx.me/protr/}
  \item \url{https://github.com/nanxstats/protr}
  \item \url{http://protr.org}
  \item Report bugs at \url{https://github.com/nanxstats/protr/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID})

Authors:
\itemize{
  \item Qing-Song Xu \email{qsxu@csu.edu.cn}
  \item Dong-Sheng Cao
}

Other contributors:
\itemize{
  \item Sebastian Mueller (Alva Genomics) [contributor]
}

}
\keyword{internal}
